requirements.txt
# Universal File Converter Requirements
# Install with: pip install -r requirements.txt

# Media processing libraries
ffmpeg-python>=0.2.0
Pillow>=10.0.0
opencv-python>=4.5.0

# Document processing
python-docx>=0.8.11
PyPDF2>=2.0.0
openpyxl>=3.0.0
python-pptx>=0.6.21

# Audio processing
pydub>=0.25.1
mutagen>=1.45.1

# Video processing
moviepy>=1.0.3

# General utilities
tqdm>=4.64.0
pathlib>=1.0.1
